requirements.txt
pdf2image
Pillow
tqdm
python-dotenv
requests
tenacity
